// src/@types/MoviesDetails.d.ts
type MovieDetailsType = {
  id: number;
  title: string;
  overview: string;
  poster_path: string;
  backdrop_path: string;
  runtime: string;
  release_date: string;
  vote_average: number;
}

export type Movie = {
  id: number;
  title: numbe;
  poster_path: string;
  runtime: string;
  release_data: string;
  vote_average: number;
}

export type RouterPropsType = {
  movieId: number;
}

export default MovieDetailsType